- cmdhist2bash: update for new format with end time and duration
these days we're logging additional fields which we were not before.
this will be tricky to transition, but we are moving to sql basis
anyways.  however we may not for the initial history load on shell
start; the raw speed of just loading a file with no callout to sqlite3
will probably be hard to beat, so we'll probably end up generating the
bash files as async+oob somehow later.  once we have implemented the sql
access as bash builtins, maybe we can look at this again.  we'll see in
time